+++ src/ccbr_actions/release.py
@@ -34,8 +34,9 @@ def prepare_draft_release(
 ):
     release_notes_filepath = path_resolve(release_notes_filepath)
     changelog_filepath = path_resolve(changelog_filepath)
-    citation_filepath = path_resolve(citation_filepath)
     version_filepath = path_resolve(version_filepath)
+    citation_filepath = path_resolve(citation_filepath)
+    assert all([f.is_file() for f in (changelog_filepath, version_filepath)])
 
     next_version = get_release_version(
         next_version_manual=next_version_manual,
@@ -56,10 +57,16 @@ def prepare_draft_release(
     write_lines(release_notes_filepath, next_release_lines, debug=debug)
     write_lines(changelog_filepath, changelog_lines, debug=debug)
     write_lines(version_filepath, [f"{next_version_strict}\n"], debug=debug)
-    update_citation(citation_file=citation_filepath, version=next_version, debug=debug)
+
+    if citation_filepath.is_file():
+        update_citation(
+            citation_file=citation_filepath, version=next_version, debug=debug
+        )
+    else:
+        citation_filepath = ""
 
     changed_files = [
-        str(f) for f in (citation_filepath, changelog_filepath, version_filepath)
+        str(f) for f in (citation_filepath, changelog_filepath, version_filepath) if f
     ]
     precommit_run(f'--files {" ".join(changed_files)}')
     push_release_draft_branch(
@@ -101,6 +108,10 @@ def post_release_cleanup(
     citation_filepath="CITATION.cff",
     debug=False,
 ):
+    changelog_filepath = path_resolve(changelog_filepath)
+    version_filepath = path_resolve(version_filepath)
+    citation_filepath = path_resolve(citation_filepath)
+
     with open(changelog_filepath, "r") as infile:
         lines = infile.readlines()
     lines.insert(0, f"## { os.path.basename(repo) } {dev_header}\n\n")
@@ -112,6 +123,7 @@ def post_release_cleanup(
         [
             str(filepath)
             for filepath in (citation_filepath, changelog_filepath, version_filepath)
+            if filepath.is_file()
         ]
     )