+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## actions development version
 
+- make citation files optional for `draft-release` and `post-release`. (#44, @kelly-sovacool)
+
 ## actions 0.2.2
 
 - fix `print_versions.py` script to work with new JSON version proposed [here](https://github.com/CCBR/Dockers2/issues/150). (#39, @kopardev)